requirements.txt
flask
flask-cors
flask-jwt-extended
python-dotenv
mysql-connector-python
gunicorn
